# docs/swagger.yaml
basePath: /
definitions:
...
  domain.ChangeReservationStatusRequest:
    properties:
      new_status:
        $ref: '#/definitions/domain.ReservationStatus'
      notes:
        type: string
    required:
    - new_status
    type: object
...
  domain.CreateTimelineEventRequest:
    properties:
      description:
        maxLength: 1000
        minLength: 1
        type: string
      title:
        maxLength: 255
        minLength: 1
        type: string
      variant:
        enum:
        - default
        - success
        - warning
        - error
        - info
        type: string
    required:
    - description
    - title
    type: object
...
  domain.DriverFeedback:
    properties:
      comment:
        type: string
      created_at:
        type: string
      id:
        type: string
      passenger_name:
        type: string
      rating:
        type: integer
      trip:
        allOf:
        - $ref: '#/definitions/domain.Reservation'
        description: Related data
      trip_id:
        type: string
    type: object
...
  domain.Payment:
    properties:
      amount:
        type: number
      created_at:
        type: string
      currency:
        type: string
      gateway:
        $ref: '#/definitions/domain.PaymentGateway'
      id:
        type: string
      payload:
        additionalProperties: true
        type: object
      reservation:
        allOf:
        - $ref: '#/definitions/domain.Reservation'
        description: Related data
      reservation_id:
        type: string
      status:
        $ref: '#/definitions/domain.PaymentStatus'
      transaction_ref:
        type: string
    type: object
  domain.PaymentGateway:
    enum:
    - WEBPAY_PLUS
    type: string
    x-enum-varnames:
    - PaymentGatewayWebpayPlus
  domain.PaymentStatus:
    enum:
    - APPROVED
    - REJECTED
    - PENDING
    type: string
    x-enum-varnames:
    - PaymentStatusApproved
    - PaymentStatusRejected
    - PaymentStatusPending
...
  domain.Reservation:
    properties:
      amount:
        type: number
      created_at:
        type: string
      datetime:
        type: string
      destination:
        type: string
      feedback:
        items:
          $ref: '#/definitions/domain.DriverFeedback'
        type: array
      id:
        description: e.g., RSV-1042
        type: string
      notes:
        type: string
      org_id:
        type: string
      passengers:
        type: integer
      payments:
        items:
          $ref: '#/definitions/domain.Payment'
        type: array
      pickup:
        type: string
      status:
        $ref: '#/definitions/domain.ReservationStatus'
      timeline:
        items:
          $ref: '#/definitions/domain.TimelineEvent'
        type: array
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/domain.User'
        description: Related data
      user_id:
        type: string
    type: object
  domain.ReservationStatus:
    enum:
    - ACTIVA
    - PROGRAMADA
    - COMPLETADA
    - CANCELADA
    type: string
    x-enum-varnames:
    - ReservationStatusActiva
    - ReservationStatusProgramada
    - ReservationStatusCompletada
    - ReservationStatusCancelada
...
  domain.TimelineEvent:
    properties:
      at:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      reservation_id:
        type: string
      title:
        type: string
      variant:
        type: string
    type: object
...
  domain.UpdateReservationRequest:
    properties:
      amount:
        type: number
      datetime:
        type: string
      destination:
        maxLength: 500
        minLength: 5
        type: string
      notes:
        maxLength: 1000
        type: string
      passengers:
        minimum: 1
        type: integer
      pickup:
        maxLength: 500
        minLength: 5
        type: string
    type: object
...
  domain.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      org_id:
        type: string
      role:
        $ref: '#/definitions/domain.UserRole'
      status:
        $ref: '#/definitions/domain.UserStatus'
      updated_at:
        type: string
    type: object
  domain.UserRole:
    enum:
    - ADMIN
    - USER
    - DRIVER
    - COMPANY
    type: string
    x-enum-varnames:
    - UserRoleAdmin
    - UserRoleUser
    - UserRoleDriver
    - UserRoleCompany
  domain.UserStatus:
    enum:
    - ACTIVE
    - BLOCKED
    type: string
    x-enum-varnames:
    - UserStatusActive
    - UserStatusBlocked
...
  domain.VehicleType:
    enum:
    - BUS
    - VAN
    - SEDAN
    - SUV
    type: string
    x-enum-varnames:
    - VehicleTypeBus
    - VehicleTypeVan
    - VehicleTypeSedan
    - VehicleTypeSUV
  handler.CreateReservationRequest:
    properties:
      datetime:
        type: string
      destination:
        maxLength: 500
        minLength: 5
        type: string
      has_special_language:
        type: boolean
      notes:
        maxLength: 1000
        type: string
      org_id:
        type: string
      passengers:
        minimum: 1
        type: integer
      pickup:
        maxLength: 500
        minLength: 5
        type: string
      stops:
        minimum: 0
        type: integer
      user_id:
        type: string
      vehicle_type:
        $ref: '#/definitions/domain.VehicleType'
    required:
    - datetime
    - destination
    - passengers
    - pickup
    - vehicle_type
    type: object
  handler.ErrorResponse:
    properties:
      details:
        type: string
      error:
        type: string
    type: object
  handler.MessageResponse:
    properties:
      message:
        type: string
    type: object
  handler.PaginatedResponse:
    properties:
      data: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@turivo.com
    name: API Support
    url: http://www.turivo.com/support
  description: API backend for Turivo transportation management system
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Turivo Backend API
  version: "1.0"
paths:
...
  /api/v1/reservations:
    get:
      consumes:
      - application/json
      description: Get paginated list of reservations
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      - description: Filter by status
        enum:
        - ACTIVA
        - PROGRAMADA
        - COMPLETADA
        - CANCELADA
        in: query
        name: status
        type: string
      - description: Filter from date (RFC3339)
        in: query
        name: from
        type: string
      - description: Filter to date (RFC3339)
        in: query
        name: to
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Sort by field
        enum:
        - datetime
        - created_at
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.Reservation'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List reservations
      tags:
      - reservations
    post:
      consumes:
      - application/json
      description: Create a new reservation with automatic price calculation
      parameters:
      - description: Reservation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateReservationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Reservation'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create reservation
      tags:
      - reservations
  /api/v1/reservations/{id}:
    get:
      consumes:
      - application/json
      description: Get reservation by ID
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Reservation'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get reservation
      tags:
      - reservations
    patch:
      consumes:
      - application/json
      description: Update reservation by ID
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: string
      - description: Reservation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateReservationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Reservation'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update reservation
      tags:
      - reservations
  /api/v1/reservations/{id}/status:
    patch:
      consumes:
      - application/json
      description: Change the status of a reservation
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: string
      - description: Status change data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.ChangeReservationStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Reservation'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Change reservation status
      tags:
      - reservations
  /api/v1/reservations/{id}/timeline:
    get:
      consumes:
      - application/json
      description: Get timeline events for a reservation
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.TimelineEvent'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get reservation timeline
      tags:
      - reservations
    post:
      consumes:
      - application/json
      description: Add a new event to reservation timeline
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: string
      - description: Timeline event data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.CreateTimelineEventRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add timeline event
      tags:
      - reservations
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"